+++ packages/abnf-rust/rust/pyo3/src/lib.rs
@@ -19,7 +19,7 @@ mod external;
 mod hooks;
 mod iter;
 mod nodes;
-mod offset;
+mod source;
 mod parsers;
 mod recursion;